renamning - use productcard in carousel
