+++ app/components/Cards/ProductsCarousel.tsx
@@ -1,0 +1,61 @@
+"use client";
+import { IProducts } from "@/app/types/products";
+import React, { useCallback, useEffect, useState } from "react";
+import {
+  Carousel,
+  CarouselContent,
+  CarouselItem,
+  CarouselNext,
+  CarouselPrevious,
+} from "@/app/components/ui/carousel";
+import { calculatePrice } from "@/app/utils/productUtils";
+import ProductCard from "./ProductCard";
+
+const ProductsCarousel = () => {
+  const [data, setData] = useState<IProducts[]>([]);
+  const [loading, setLoading] = useState(true);
+
+  const getData = useCallback(async () => {
+    try {
+      const response = await fetch("/api/getLimitedProducts");
+      if (!response.ok) {
+        throw new Error("Failed to fetch products");
+      }
+      const productsData = await response.json();
+      setData(productsData);
+      setLoading(false);
+    } catch (error) {
+      console.error(error);
+      setLoading(false);
+    }
+  }, []);
+
+  useEffect(() => {
+    getData();
+  }, [getData]);
+
+  if (loading) return <div>Loading...</div>;
+  const firstSevenProducts = data.slice(0, 7);
+  const previewProductsData = calculatePrice(firstSevenProducts);
+
+  return (
+    <Carousel
+      opts={{
+        align: "start",
+      }}
+      className="w-[90%]"
+    >
+      <CarouselContent>
+        {previewProductsData.map((product) => (
+          <CarouselItem key={product.id}>
+            <ProductCard key={product.id} {...product} />
+          </CarouselItem>
+        ))}
+      </CarouselContent>
+      <CarouselPrevious />
+      <CarouselNext />
+    </Carousel>
+  );
+};
+
+export default ProductsCarousel;
