+++ app/page.tsx
@@ -1,12 +1,12 @@
 import React from "react";
-import PreviewProductsCards from "./components/Cards/PreviewCards";
+import ProductsCarousel from "./components/Cards/ProductsCarousel";
 import HeroBanner from "./components/HeroBanner/HeroBanner";
 
 export default function Home() {
   return (
     <>
       <HeroBanner />
-      <PreviewProductsCards />
+      <ProductsCarousel />
     </>
   );
 }
